+++ example/mit_ps.conf
@@ -51,5 +51,5 @@ debug = true
 # length of factor
 emedding_size = 4
 field_num = 10
-max_key = 1000000
+max_key = 10000000
 num_thread = 3
