+++ openmit/cli_main.cc
@@ -9,7 +9,10 @@ int main(int argc, char * argv[]) {
   if (strcmp(argv[1], "none")) parser.ReadFile(argv[1]);
   parser.ReadArgs(argc - 2, argv + 2);
   const mit::KWArgs kwargs = parser.GetKWArgs();
-  std::unique_ptr<mit::MILearner> milearner(mit::MILearner::Create(kwargs)); 
+
+  mit::MILearner* milearner = mit::MILearner::Create(kwargs);
   milearner->Run();
+  delete milearner; 
+  milearner = NULL;
   return 0;
 }
