+++ openmit/common/base.h
@@ -1,7 +1,7 @@
 /*!
  *  Copyright (c) 2016 by Contributors
  *  \file base.h
- *  \brief define macros
+ *  \brief define macros and const variables
  *  \author ZhouYong
  */
 #ifndef OPENMIT_COMMON_BASE_H_
@@ -11,15 +11,14 @@
 
 namespace mit {
 /*!
- * \brief unsigned integer type used in mit,
- *        used for feature index, field index and row index.
+ * \brief unsigned integer type used in openmit
  */
 typedef uint64_t mit_uint;
 
-/*! \brief float used for weight and calculation values. */
+/*! \brief type real used in openmit */
 typedef float mit_float;
 
-inline mit_uint NewKey(mit_uint featid, mit_uint fieldid, size_t nbit) {
+inline mit_uint NewKey(mit_uint featid, size_t fieldid, size_t nbit) {
   return (featid << nbit) + fieldid;
 }
