+++ openmit/learner/mi_learner.h
@@ -1,13 +1,13 @@
 /*!
  *  Copyright (c) 2016 by Contributors
  *  \file mi_learner.h
- *  \brief Machine Intelligence Toolkits Learner
+ *  \brief machine intelligence toolkits learner
  *  \author ZhouYong
  */
 #ifndef OPENMIT_LEARNER_MI_LEARNER_H_
 #define OPENMIT_LEARNER_MI_LEARNER_H_
 
-#include <memory>
+//#include <memory>
 #include <string>
 #include "dmlc/parameter.h"
 #include "openmit/common/arg.h"
@@ -19,17 +19,18 @@ namespace mit {
  */
 class MILearner {
   public:
-    /*! create learner by conf args info */
-    static MILearner * Create(const mit::KWArgs & kwargs);
+    /*! brief create learner by conf args info */
+    static MILearner * Create(const mit::KWArgs& kwargs);
     
-    /*! constructor */
+    /*! brief constructor */
     MILearner() {}
     
-    /*! destructor */
+    /*! brief destructor */
     virtual ~MILearner() {}
 
-    /*! running */
+    /*! brief running machine intelligence task */
     virtual void Run() = 0;
 }; // class MITLearner
+
 } // namespace mit
 #endif // OPENMIT_LEARNER_MI_LEARNER_H_
