+++ openmit/models/model.h
@@ -10,9 +10,8 @@
 #include "dmlc/logging.h"
 #include "openmit/common/arg.h"
 #include "openmit/common/base.h"
-#include "openmit/common/data/data.h"
-#include "openmit/common/parameter/cli_param.h"
-#include "openmit/common/parameter/model_param.h"
+#include "openmit/common/data.h"
+#include "openmit/common/parameter.h"
 #include "openmit/entity/entry_meta.h"
 #include "openmit/optimizer/optimizer.h"
 #include "openmit/tools/math/basic_formula.h"
